+++ tmtccmd/pus/s5_event.py
@@ -1,3 +1,0 @@
-from .s5_event_defs import *  # noqa re-export
-from tmtccmd.tm.pus_5_fsfw_event import *  # noqa re-export
-from tmtccmd.tc.pus_5_event import *  # noqa re-export