+++ tmtccmd/pus/s5_satrs_event.py
@@ -1,0 +1,3 @@
+"""`sat-rs <https://egit.irs.uni-stuttgart.de/rust/sat-rs.git>`_ specific PUS event support."""
+from .s5_satrs_event_defs import *  # noqa re-export
+from tmtccmd.tc.pus_5_event import *  # noqa re-export